+++ src/components/Header/Header.test.tsx
@@ -63,7 +63,7 @@ describe('Header', () => {
     render(headerComponent());
     const navItemTrigger = screen.getByTestId(`nav-item-trigger-auctions`);
     const backBtn = screen.getByText(`Back`);
-    const navLabel = screen.getByTestId('phillips-main-nav-label');
+    const navLabel = screen.getByTestId(`${px}-main-nav-label`);
     await userEvent.click(navItemTrigger);
     expect(navItemTrigger).toHaveClass(`${px}-nav__item--expanded`);
     expect(navLabel).toHaveTextContent('Auctions');